submission/requirements.txt
# Requirements for competition submission
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.56.0
Pillow>=9.4.0
numpy>=1.24.0
